third_party/blink/web_tests/dom/mutation-event-tests/dom/node/DOMNodeRemovedEvent-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser very soon. Usage of this event listener will cause performance issues today, and represents a large risk of future site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
This test checks that DOMNodeRemovedEvent is emitted once (and only once).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS count is 1
PASS successfullyParsed is true

TEST COMPLETE
